+++ civirules/conditions.mgd.php
@@ -7,5 +7,5 @@ use CRM_CampagnodonCivicrm_ExtensionUtil as E;
 require_once(E::path('civirules/civirules_installed.php'));
 
 if (_campagnodon_is_civirules_installed()) {
-  CRM_Civirules_Utils_Upgrader::insertTriggersFromJson(E::path('civirules/triggers.json'));
+  CRM_Civirules_Utils_Upgrader::insertConditionsFromJson(E::path('civirules/conditions.json'));
 }